fix: resolve design issue with 2.0.0 being unable to kill process
Resolve a design issue in 2.0.0 where the process that was spawned could
not be killed. This happened because we started _only_ returning the
promise, which meant that things worked fine for processes that
terminated on their own, but not for processes that never terminated
on their own. Such as a server that needs to be explicitly killed.

Signed-off-by: David Piper <david.piper@mailfence.com>